change implementation to not require Comparable conformance for ActorID, as it causes too much pain with adding retroactive conformance to a broad protocol
