+++ Package.swift
@@ -24,7 +24,19 @@ let package = Package(
     targets: [
         .target(
             name: "CRDT",
-            dependencies: []
+            dependencies: [],
+            swiftSettings: [
+                .unsafeFlags(["-Xfrontend", "-strict-concurrency=complete"]),
+                /*
+                    Summation from https://www.donnywals.com/enabling-concurrency-warnings-in-xcode-14/
+                    Set `strict-concurrency` to `targeted` to enforce Sendable and actor-isolation checks in your code.
+                      This explicitly verifies that `Sendable` constraints are met when you mark one of your types as `Sendable`.
+                      This mode is essentially a bit of a hybrid between the behavior that's intended in Swift 6, and the default in Swift 5.7.
+                      Use this mode to have a bit of checking on your code that uses Swift concurrency without too many warnings and / or errors in your current codebase.
+
+                    Set `strict-concurrency` to `complete` to get the full suite of concurrency constraints, essentially as they will work in Swift 6.
+                    */
+            ]
         ),
         .testTarget(
             name: "CRDTTests",
