+++ Sources/CRDT/GenericCRDTs/LWWRegister.swift
@@ -12,7 +12,7 @@ import Foundation
 /// The `LWWRegister` implementation is based on an optimized CRDT register type as described in
 /// [A comprehensive study of Convergent and Commutative Replicated Data Types](https://hal.inria.fr/inria-00555588/document)”
 /// by Marc Shapiro, Nuno Preguiça, Carlos Baquero, and Marek Zawirski (2011).
-public struct LWWRegister<ActorID: Hashable & Comparable, T> {
+public struct LWWRegister<ActorID: Hashable & PartiallyOrderable, T> {
     /// A struct that represents the state of an LWWRegister
     public struct Metadata {
         internal var value: T
