+++ Sources/CRDT/GenericCRDTs/PNCounter.swift
@@ -7,7 +7,7 @@
 /// The `PNCounter` implementation is based on the PN Counter as described
 /// [A comprehensive study of Convergent and Commutative Replicated Data Types](https://hal.inria.fr/inria-00555588/document)”
 /// by Marc Shapiro, Nuno Preguiça, Carlos Baquero, and Marek Zawirski (2011).
-public struct PNCounter<ActorID: Hashable & Comparable> {
+public struct PNCounter<ActorID: Hashable & PartiallyOrderable> {
     internal var pos_value: UInt
     internal var neg_value: UInt
     internal let selfId: ActorID
