+++ Tests/CRDTTests/LamportTimestampTests.swift
@@ -25,15 +25,15 @@ final class LamportTimestampTests: XCTestCase {
         var ts1 = LamportTimestamp(actorId: 1)
         var ts2 = LamportTimestamp(actorId: 2)
         // partial ordering is by count FIRST, then by ID if equal
-        XCTAssertTrue(ts1 < ts2)
+        XCTAssertTrue(ts1 <= ts2)
         XCTAssertEqual(ts1.clock, ts2.clock)
 
         ts1.tick()
         XCTAssertTrue(ts1.clock > ts2.clock)
-        XCTAssertFalse(ts1 < ts2)
+        XCTAssertFalse(ts1 <= ts2)
 
         ts2.tick()
-        XCTAssertTrue(ts1 < ts2)
+        XCTAssertTrue(ts1 <= ts2)
         XCTAssertEqual(ts1.clock, ts2.clock)
     }
 }
